chore: split PicData structs into separate file
