+++ src/pic/data/grayscale.rs
@@ -1,0 +1,8 @@
+use crate::pic::coordinatesystem::CoordinateSystem;
+use crate::parser::aptnode::APTNode;
+
+#[derive(Clone, Debug, PartialEq)]
+pub struct GrayscaleData {
+    pub c: APTNode,
+    pub coord: CoordinateSystem,
+}
