+++ src/pic/data/mod.rs
@@ -1,0 +1,11 @@
+pub mod gradient;
+pub mod mono;
+pub mod grayscale;
+pub mod rgb;
+pub mod hsv;
+
+pub use mono::MonoData;
+pub use grayscale::GrayscaleData;
+pub use rgb::RGBData;
+pub use hsv::HSVData;
+pub use gradient::GradientData;
