+++ src/pic/data/mono.rs
@@ -1,0 +1,8 @@
+use crate::pic::coordinatesystem::CoordinateSystem;
+use crate::parser::aptnode::APTNode;
+
+#[derive(Clone, Debug, PartialEq)]
+pub struct MonoData {
+    pub c: APTNode,
+    pub coord: CoordinateSystem,
+}
